Refactor the reference manager to a new module.
